+++ betterocr/engines/easy_pororo_ocr/pororo/models/brainOCR/detection.py
