+++ betterocr/engines/easy_pororo_ocr/pororo/tasks/optical_character_recognition.py
