+++ betterocr/engines/easy_pororo_ocr/pororo/tasks/utils/base.py
