+++ betterocr/ocr_wrappers/tesseract/job.py
