+++ betterocr/ocr_wrappers/tesseract/mapping.py
